requirements.txt
# Flask Core
Flask==3.0.3
Flask-SQLAlchemy==3.1.1

python-dotenv==1.0.1
Flask-Login==0.6.3
Werkzeug==3.1.3

# Security & Validation
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Cors==4.0.0
pydantic==2.11.7
validators==0.22.0
cryptography==42.0.8

# Real-time features
Flask-SocketIO==5.4.1

# Payment processing - Indian Gateways
razorpay==1.3.0
paypalrestsdk==1.13.3
requests==2.32.5

# Task Queue & Background Jobs
celery==5.3.6
redis==4.6.0

# AI Chatbot
openai==1.51.2
google-generativeai==0.8.3

# Multi-factor authentication
pyotp==2.9.0

# Email notifications
sendgrid==6.11.0
Flask-Mail==0.10.0

# SMS notifications
twilio==9.2.4

# QR Code generation
qrcode==7.4.2
Pillow==10.4.0

# PDF Generation for E-Tickets
reportlab==4.4.3

# Admin Dashboard
Flask-Admin==1.6.1

# Additional utilities
jinja2==3.1.6
markupsafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
rich==13.9.4
ordered-set==4.1.0
limits==5.5.0
typing-extensions==4.15.0
wtforms==3.2.1

# Production WSGI Server
gunicorn==22.0.0

# Monitoring & Logging
sentry-sdk==2.17.0